Change requestId variable type to quint64 from int.
Took 6 minutes

+++ cockatrice/src/interface/widgets/deck_editor/deck_editor_deck_dock_widget.h
@@ -103,7 +103,7 @@ private:
 
     BracketExplanation lastBracketExplanation;
 
-    int requestId;
+    quint64 requestId;
 
     DeckListModel *getModel() const;
     [[nodiscard]] QModelIndexList getSelectedCardNodeSourceIndices() const;
